Rework SetVideoParams a little - no real need to go through all that if just the keyframedist is changing. Qt 3.1 compile fix for tsstats.h.
git-svn-id: http://svn.mythtv.org/svn/trunk@5236 7dbf422c-18fa-0310-86e9-fd20926502f2

+++ mythtv/libs/libmythtv/mpeg/tsstats.h
@@ -22,11 +22,11 @@ class TSStats
 
 inline QString TSStats::toString() {
     QString str("Transport Stream Statistics\n");
-    str.append(QString("TSPacket Count: %1").arg(_tspacket_count));
+    str.append(QString("TSPacket Count: %1").arg((long)_tspacket_count));
     QMapIterator<int, long long> it = _pid_counts.begin();
     for (; it != _pid_counts.end(); it++)
         str.append(QString("\nPID 0x%1 Count: %2").
-                   arg((int)it.key(),0,16).arg((long long)it.data(),10,10));
+                   arg((int)it.key(),0,16).arg((long)it.data(),10,10));
     return str;
 }
